Add pg scopes as recommended by OTP
The default scope is there only for easier migration from the old pg2,
but what is advised is to define specific scopes for each process group,
see:

https://www.erlang.org/doc/man/pg.html
> Process groups can be organised into multiple scopes. Scopes are
completely independent of each other. A process may join any number of
groups in any number of scopes. Scopes are designed to decouple single
mesh into a set of overlay networks, reducing amount of traffic required
to propagate group membership information. Default scope pg is started
automatically when kernel(6) is configured to do so.